requirements.txt
affine==2.3.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextily==1.0.1
cycler==0.10.0
descartes==1.1.0
Fiona==1.8.18
geographiclib==1.50
geopandas==0.8.1
geopy==2.1.0
idna==2.10
joblib==1.0.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.3
mercantile==1.1.6
munch==2.5.0
neworder==0.2.1
numpy==1.19.5
pandas==1.2.0
Pillow==8.1.0
police-api-client==1.2.2
pykml==0.2.0
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.5
rasterio==1.1.8
requests==2.25.1
responses==0.12.1
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
urllib3==1.26.2
